requirements.txt
protobuf>=3.19.5,<3.20.1
transformers==4.26.1
icetk
cpm_kernels
torch>=1.10
python-telegram-bot>=20.1